requirements.txt
Django==5.2.13
firebase-admin==7.1.0
python-dotenv==1.1.1
pyserial==3.5
